Complete action and reducers set up
