--- src/features/books/booksSlice.js
@@ -1,19 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-const initialState = {
-  books: [],
-};
-
-const booksSlice = createSlice({
-  name: 'books',
-  initialState,
-  reducers: {
-    addBook: (state, action) => {
-      state.books.push(action.payload);
-    },
-    removeBook: (state, action) => {
-      const { id } = action.payload;
-      state.books.filter((book) => book.id !== id);
-    },
-  },
-});
