+++ src/redux/categories/categoriesSlice.js
@@ -1,0 +1,14 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState = {
+  categories: [],
+};
+const categoriesSlice = createSlice({
+  name: 'categories',
+  initialState,
+  reducers: {
+    checkStatus: (state) => state.categories.length && 'Under construction',
+  },
+});
+
+export default categoriesSlice.reducer;
