+++ src/redux/store.js
@@ -1,9 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from '../features/books/booksSlice';
+import categoriesReducer from './categories/categoriesSlice';
+import booksReducer from './books/booksSlice';
 
 const store = configureStore({
   reducer: {
     books: booksReducer,
+    categories: categoriesReducer,
   },
 });
